pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stix2arango"
version = "0.0.1"
authors = [
  { name="DOGESEC", email="support@dogesec.com" },
]
description = "A package containing the stix definitions for custom objects used at dogesec"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "stix2",
    "python-arango==7.9.1; python_version >= '3.8'",
    "tqdm==4.66.4; python_version >= '3.7'",
    "jsonschema==4.22.0; python_version >= '3.8'",
    "requests==2.31.0; python_version >= '3.7'",
  ]
[project.urls]
Homepage = "https://github.com/muchdogesec/stix2arango"
Issues = "https://github.com/muchdogesec/stix2arango/issues"

[project.scripts]
stix2arango = "stix2arango.__main__:main"
